Refs #221 Unfinished Restructuring, integrating Loader for cluster and instance configuration(s)
