+++ bibigrid-core/src/main/java/de/unibi/cebitec/bibigrid/core/intents/ListIntent.java
@@ -10,180 +10,21 @@ import java.util.*;
 /**
  * Creates a Map of BiBiGrid cluster instances
  * TODO clusterMap only accessible via ListIntent ...
+ *      Separate list intent from load config and instance values
  *
  * @author Johannes Steiner - jsteiner(at)cebitec.uni-bielefeld.de
  * @author Jan Krueger - jkrueger(at)cebitec.uni-bielefeld.de
  */
 public abstract class ListIntent extends Intent {
     private static final Logger LOG = LoggerFactory.getLogger(ListIntent.class);
-    private static final DateTimeFormatter dateTimeFormatter = DateTimeFormatter.ofPattern("dd/MM/yy HH:mm:ss");
 
-    // TODO module used to get ListIntent. Probably decouple it here, only used for instanceType
-    protected final ProviderModule providerModule;
-    protected final Client client;
-    protected final Configuration config;
     private Map<String, Cluster> clusterMap;
 
-    protected ListIntent(ProviderModule providerModule, Client client, Configuration config) {
-        this.providerModule = providerModule;
-        this.client = client;
-        this.config = config;
+    protected ListIntent(Map<String, Cluster> clusterMap) {
+        this.clusterMap = clusterMap;
     }
 
-    /**
-     * Extract cluster from clusterMap or create a new one if not yet existent.
-     * TODO little confusing since the class CreateCluster is not used here.
-     *      Instead ListIntent uses only Cluster as representation
-     * @param clusterId Id of cluster
-     * @return cluster
-     */
-    protected final Cluster getOrCreateCluster(String clusterId) {
-        Cluster cluster;
-        if (clusterMap.containsKey(clusterId)) {
-            cluster = clusterMap.get(clusterId);
-        } else {
-            cluster = new Cluster(clusterId);
-            clusterMap.put(clusterId, cluster);
-        }
-        return cluster;
-    }
 
-    /**
-     * Return a Map of Cluster objects within current configuration.
-     */
-    public final Map<String, Cluster> getList() {
-        if (clusterMap == null) {
-            clusterMap = new HashMap<>();
-            assignClusterConfigValues();
-        }
-        return clusterMap;
-    }
-
-    /**
-     * Extracts cluster Id from name, e.g. bibigrid-worker1-2-jbxoecbvtfu3n5f.
-     * TODO Probably an error source, if names change -> search in clusterMap and getId ?
-     * @param name instance name
-     * @return clusterId
-     */
-    protected static String getClusterIdFromName(String name) {
-        String[] parts = name.split("-");
-        return parts[parts.length - 1];
-    }
-
-    /**
-     * Assigns client and cluster values from single servers to internal config.
-     */
-    protected void assignClusterConfigValues() {
-        List<Instance> instances = getInstances();
-        if (instances != null) {
-            for (Instance instance : instances) {
-                checkInstance(instance);
-            }
-        }
-        List<Network> networks = client.getNetworks();
-        if (networks != null) {
-            for (Network network : networks) {
-                String name = network.getName();
-                if (name != null && name.startsWith(CreateClusterEnvironment.NETWORK_PREFIX)) {
-
-                    getOrCreateCluster(getClusterIdFromName(name)).setNetwork(network);
-                }
-            }
-        }
-        List<Subnet> subnets = client.getSubnets();
-        if (subnets != null) {
-            for (Subnet subnet : subnets) {
-                String name = subnet.getName();
-                if (name != null && name.startsWith(CreateClusterEnvironment.SUBNET_PREFIX)) {
-                    getOrCreateCluster(getClusterIdFromName(name)).setSubnet(subnet);
-                }
-            }
-        }
-        List<String> keypairs = client.getKeypairNames();
-        if (keypairs != null) {
-            for (String name : keypairs) {
-                if (name != null && name.startsWith(CreateCluster.PREFIX)) {
-                    getOrCreateCluster(getClusterIdFromName(name)).setKeyName(name);
-                }
-            }
-        }
-    }
-
-    protected abstract List<Instance> getInstances();
-
-    /**
-     * Loads instance configurations from internal and sets cluster and instance config.
-     * @param instance master or worker
-     */
-    protected void checkInstance(Instance instance) {
-        // check if instance is a BiBiGrid instance and extract clusterId from it
-        String clusterId = getClusterIdForInstance(instance);
-        if (clusterId == null)
-            return;
-        Cluster cluster = getOrCreateCluster(clusterId);
-        loadInstanceConfiguration(instance);
-        // Check whether master or worker instance
-        String name = instance.getTag(Instance.TAG_NAME);
-        if (name == null) {
-            name = instance.getName();
-        }
-        if (name != null && name.startsWith(CreateCluster.MASTER_NAME_PREFIX)) {
-            if (cluster.getMasterInstance() == null) {
-                cluster.setMasterInstance(instance);
-                cluster.setPublicIp(instance.getPublicIp());
-                cluster.setPrivateIp(instance.getPrivateIp());
-                cluster.setKeyName(instance.getKeyName());
-                cluster.setStarted(instance.getCreationTimestamp().format(dateTimeFormatter));
-            } else {
-                LOG.error("Detected two master instances ({},{}) for cluster '{}'.", cluster.getMasterInstance().getName(),
-                        instance.getName(), clusterId);
-            }
-        } else {
-            cluster.addWorkerInstance(instance);
-        }
-        checkInstanceKeyName(instance, cluster);
-        checkInstanceUserTag(instance, cluster);
-    }
-
-    private void checkInstanceKeyName(Instance instance, Cluster cluster) {
-        //key name should be always the same for all instances of one cluster
-        if (cluster.getKeyName() != null) {
-            if (!cluster.getKeyName().equals(instance.getKeyName())) {
-                LOG.error("Detected two different keynames ({},{}) for cluster '{}'.", cluster.getKeyName(),
-                        instance.getKeyName(), cluster.getClusterId());
-            }
-        } else {
-            cluster.setKeyName(instance.getKeyName());
-        }
-    }
-
-    private void checkInstanceUserTag(Instance instance, Cluster cluster) {
-        // user should be always the same for all instances of one cluster
-        String user = instance.getTag(Instance.TAG_USER);
-        if (user != null) {
-            if (cluster.getUser() == null) {
-                cluster.setUser(user);
-            } else if (!cluster.getUser().equals(user)) {
-                LOG.error("Detected two different users ({},{}) for cluster '{}'.", cluster.getUser(), user,
-                        cluster.getClusterId());
-            }
-        }
-    }
-
-    private String getClusterIdForInstance(Instance instance) {
-        String clusterIdTag = instance.getTag(Instance.TAG_BIBIGRID_ID);
-        if (clusterIdTag != null) {
-            return clusterIdTag;
-        }
-        String name = instance.getName();
-        if (name != null && (name.startsWith(CreateCluster.MASTER_NAME_PREFIX) ||
-                name.startsWith(CreateCluster.WORKER_NAME_PREFIX))) {
-            return getClusterIdFromName(name);
-        }
-        return null;
-    }
-
-    protected abstract void loadInstanceConfiguration(Instance instance);
 
     /**
      * Return a String representation of found cluster objects map.
@@ -191,13 +32,6 @@ public abstract class ListIntent extends Intent {
      */
     @Override
     public final String toString() {
-        if (clusterMap == null) {
-            clusterMap = new HashMap<>();
-            assignClusterConfigValues();
-        }
-        if (clusterMap.isEmpty()) {
-            return "No BiBiGrid cluster found!\n";
-        }
         StringBuilder display = new StringBuilder();
         Formatter formatter = new Formatter(display, Locale.US);
         display.append("\n");
@@ -273,25 +107,6 @@ public abstract class ListIntent extends Intent {
     }
 
     /**
-     * Determines clusterMap if not available and gets cluster with specified clusterId.
-     * @param clusterId Id of cluster
-     * @return Cluster initialized or already available in clusterMap
-     */
-    public Cluster getCluster(String clusterId) {
-        if (clusterMap == null) {
-            clusterMap = new HashMap<>();
-            assignClusterConfigValues();
-        }
-        if (clusterMap.isEmpty()) {
-            LOG.error("No BiBiGrid cluster found!\n");
-        }
-        if (!clusterMap.containsKey(clusterId)) {
-            LOG.error("No BiBiGrid cluster with id '" + clusterId + "' found!\n");
-        }
-        return clusterMap.get(clusterId);
-    }
-
-    /**
      * Displays detailed information about cluster nodes when using '-l <cluster-id>'.
      * @param display cluster node information String
      * @param instance cluster node
@@ -311,6 +126,6 @@ public abstract class ListIntent extends Intent {
             display.append("  type: ").append(instanceConfig.getType()).append("\n");
             display.append("  image: ").append(instanceConfig.getImage()).append("\n");
         }
-        display.append("  launch-date: ").append(instance.getCreationTimestamp().format(dateTimeFormatter)).append("\n");
+        display.append("  launch-date: ").append(instance.getCreationTimestamp().format(LoadClusterConfigurationIntent.DATE_TIME_FORMATTER)).append("\n");
     }
 }
